+++ src/components/Queue/index.tsx
@@ -1,27 +1,22 @@
-import { useRef } from 'react'
 import { useCurrentIndex, usePlayQueue, useQueueRef } from '../../stores/player/queue'
 import { TrackItem } from 'react-native-nitro-player'
-import { Easing, ListRenderItemInfo, StyleSheet } from 'react-native'
+import { ListRenderItemInfo, StyleSheet } from 'react-native'
 import { reorderQueue } from '../../hooks/player/functions/queue'
-import { useSafeAreaFrame, useSafeAreaInsets } from 'react-native-safe-area-context'
-import { DraxList, DraxProvider, SortableReorderEvent } from 'react-native-drax'
+import { useSafeAreaInsets } from 'react-native-safe-area-context'
+import { DraxList, SortableReorderEvent } from 'react-native-drax'
 import QueuedTrack from './components/track'
 import { itemDraxViewProps } from '../../configs/styling/drax'
-import { LegendList, LegendListRef } from '@legendapp/list/react-native'
+import { LegendList } from '@legendapp/list/react-native'
 import { FadeOut } from 'react-native-reanimated'
-import { Player } from '../Player'
 
 export default function Queue(): React.JSX.Element {
-	const { height } = useSafeAreaFrame()
-	const { bottom } = useSafeAreaInsets()
+	const { top, bottom } = useSafeAreaInsets()
 
 	const queue = usePlayQueue()
 
 	const currentIndex = useCurrentIndex()
 
 	const queueRef = useQueueRef()
-
-	const listRef = useRef<LegendListRef>(null)
 
 	const keyExtractor = (item: TrackItem) => `${item.id}`
 
@@ -36,29 +31,20 @@ export default function Queue(): React.JSX.Element {
 		<QueuedTrack {...props} queueRef={queueRef} queueIndex={queue.indexOf(props.item)} />
 	)
 
-	const scrollToCurrentTrack = () => {
-		if (currentIndex === undefined || currentIndex === null) return
-
-		listRef.current?.scrollToIndex({
-			animated: true,
-			index: currentIndex,
-		})
-	}
-
 	return (
 		<DraxList<TrackItem>
 			component={LegendList}
 			animationConfig={'spring'}
 			containerStyle={styles.container}
 			contentContainerStyle={{
+				marginTop: top,
 				marginBottom: bottom,
 			}}
 			data={queue}
 			keyExtractor={keyExtractor}
-			ref={listRef}
 			renderItem={renderItem}
 			onReorder={onReorder}
-			onLayout={scrollToCurrentTrack}
+			initialScrollIndex={currentIndex}
 			itemDraxViewProps={itemDraxViewProps}
 			lockToMainAxis
 			itemExiting={FadeOut.springify()}
@@ -69,5 +55,7 @@ export default function Queue(): React.JSX.Element {
 const styles = StyleSheet.create({
 	container: {
 		flex: 1,
+		width: '100%',
+		height: '100%',
 	},
 })
